Fixing issue of un-necessary renders
