+++ src/FlatList.tsx
@@ -1,6 +1,11 @@
-import React, { useRef, useState } from 'react';
-import { FlatList, FlatListProps } from 'react-native';
-import { useMvcpTuner } from './useMvcpTuner';
+import React, { useRef } from 'react';
+import {
+  FlatList,
+  FlatListProps,
+  findNodeHandle,
+  NativeModules,
+  Platform,
+} from 'react-native';
 
 export type maintainVisibleContentPositionPropType = {
   autoscrollToTopThreshold?: number;
@@ -10,21 +15,54 @@ export type maintainVisibleContentPositionPropType = {
 export type FlatListComponentPropType<T = any> = FlatListProps<T> & {
   maintainVisibleContentPosition: maintainVisibleContentPositionPropType;
 };
+export const MvcpScrollViewManager = NativeModules.MvcpScrollViewManager;
 
 export default React.forwardRef(
-  (props: FlatListComponentPropType, forwardedRef) => {
-    const flRef = useRef<FlatList>();
-    const [refReady, setRefReady] = useState(false);
-    const { extraData, maintainVisibleContentPosition: mvcp } = props;
-    useMvcpTuner(flRef, refReady, mvcp, extraData);
+  <T extends any>(props: FlatListComponentPropType, forwardedRef) => {
+    const flRef = useRef<FlatList<T>>();
+    const { maintainVisibleContentPosition: mvcp } = props;
+
+    const autoscrollToTopThreshold = useRef<number>();
+    const minIndexForVisible = useRef<number>();
+    const cleanupPromiseRef = useRef<Promise<any>>();
+
+    const resetMvcpIfNeeded = (): void => {
+      if (!mvcp || Platform.OS !== 'android' || !flRef?.current) {
+        return;
+      }
+      if (
+        autoscrollToTopThreshold.current === mvcp.autoscrollToTopThreshold &&
+        minIndexForVisible.current === mvcp.minIndexForVisible
+      ) {
+        // Don't do anythinig if the values haven't changed
+        return;
+      }
+
+      cleanupPromiseRef.current &&
+        cleanupPromiseRef.current?.then((handle) => {
+          MvcpScrollViewManager.disableMaintainVisibleContentPosition(handle);
+        });
+
+      autoscrollToTopThreshold.current =
+        mvcp.autoscrollToTopThreshold || -Number.MAX_SAFE_INTEGER;
+      minIndexForVisible.current = mvcp.minIndexForVisible || 0;
+
+      const viewTag = findNodeHandle(flRef.current);
+      cleanupPromiseRef.current = MvcpScrollViewManager.enableMaintainVisibleContentPosition(
+        viewTag,
+        autoscrollToTopThreshold.current,
+        minIndexForVisible.current
+      );
+    };
 
     return (
-      <FlatList
+      <FlatList<T>
         {...props}
         ref={(ref) => {
           // @ts-ignore
           flRef.current = ref;
-          setRefReady(true);
+
+          resetMvcpIfNeeded();
 
           if (typeof forwardedRef === 'function') {
             forwardedRef(ref);
