+++ src/ScrollView.tsx
@@ -1,13 +1,11 @@
-import React, { useState, useRef } from 'react';
+import React, { useRef } from 'react';
 import {
-  NativeModules,
-  FlatList,
-  ScrollViewProps,
   ScrollView,
+  ScrollViewProps,
+  findNodeHandle,
+  NativeModules,
+  Platform,
 } from 'react-native';
-import { useMvcpTuner } from './useMvcpTuner';
-
-export const MvcpScrollViewManager = NativeModules.MvcpScrollViewManager;
 
 export type maintainVisibleContentPositionPropType = {
   autoscrollToTopThreshold?: number;
@@ -17,14 +15,45 @@ export type maintainVisibleContentPositionPropType = {
 export type ScrollViewComponentPropType = ScrollViewProps & {
   maintainVisibleContentPosition: maintainVisibleContentPositionPropType;
 };
+export const MvcpScrollViewManager = NativeModules.MvcpScrollViewManager;
 
 export default React.forwardRef(
   (props: ScrollViewComponentPropType, forwardedRef) => {
-    const flRef = useRef<FlatList>();
-    const [refReady, setRefReady] = useState(false);
-
+    const flRef = useRef<ScrollView>();
     const { maintainVisibleContentPosition: mvcp } = props;
-    useMvcpTuner(flRef, refReady, mvcp);
+
+    const autoscrollToTopThreshold = useRef<number>();
+    const minIndexForVisible = useRef<number>();
+    const cleanupPromiseRef = useRef<Promise<any>>();
+
+    const resetMvcpIfNeeded = (): void => {
+      if (!mvcp || Platform.OS !== 'android' || !flRef?.current) {
+        return;
+      }
+      if (
+        autoscrollToTopThreshold.current === mvcp.autoscrollToTopThreshold &&
+        minIndexForVisible.current === mvcp.minIndexForVisible
+      ) {
+        // Don't do anythinig if the values haven't changed
+        return;
+      }
+
+      cleanupPromiseRef.current &&
+        cleanupPromiseRef.current?.then((handle) => {
+          MvcpScrollViewManager.disableMaintainVisibleContentPosition(handle);
+        });
+
+      autoscrollToTopThreshold.current =
+        mvcp.autoscrollToTopThreshold || -Number.MAX_SAFE_INTEGER;
+      minIndexForVisible.current = mvcp.minIndexForVisible || 0;
+
+      const viewTag = findNodeHandle(flRef.current);
+      cleanupPromiseRef.current = MvcpScrollViewManager.enableMaintainVisibleContentPosition(
+        viewTag,
+        autoscrollToTopThreshold.current,
+        minIndexForVisible.current
+      );
+    };
 
     return (
       <ScrollView
@@ -32,7 +61,8 @@ export default React.forwardRef(
         ref={(ref) => {
           // @ts-ignore
           flRef.current = ref;
-          setRefReady(true);
+
+          resetMvcpIfNeeded();
 
           if (typeof forwardedRef === 'function') {
             forwardedRef(ref);
